pyproject.toml
[project]
name = "career-mentor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.8",
    "openai-agents>=0.2.5",
    "pydantic>=2.11.7",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.14.1",
]
